feat: some code cleanup
